requirements.txt
Django==5.0
djangorestframework==3.16.0
django-cors-headers==4.3.1
gunicorn==23.0.0
whitenoise==6.6.0
Pillow==10.1.0
numpy==1.24.3
tflite-runtime==2.14.0
mediapipe==0.10.14
opencv-python-headless==4.8.1.78